dynamic fee - oraichain testnet

+++ packages/stores/src/common/utils/index.ts
@@ -34,12 +34,16 @@ export class StoreUtils {
     bals: CoinPrimitive[]
   ): CoinPretty[] {
     const result: CoinPretty[] = [];
-    for (const bal of bals) {
-      const currency = currenciesMap[bal.denom];
-      if (currency) {
-        const amount = new Dec(bal.amount);
-        if (amount.truncate().gt(new Int(0))) {
-          result.push(new CoinPretty(currency, amount));
+    console.log("bals", bals);
+
+    if (bals) {
+      for (const bal of bals) {
+        const currency = currenciesMap[bal.denom];
+        if (currency) {
+          const amount = new Dec(bal.amount);
+          if (amount.truncate().gt(new Int(0))) {
+            result.push(new CoinPretty(currency, amount));
+          }
         }
       }
     }
